# prooflink-master/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: ProofLink API
  version: "1.0.0"
  description: >-
    ProofLink compliance, identity, and invoicing API for autonomous agent
    payments. Provides sanctions screening, KYA credential issuance, invoice
    lifecycle management, webhook delivery, and platform analytics.
  contact:
    name: ProofLink
    url: https://prooflink.io
  license:
    name: MIT
servers:
  - url: "{baseUrl}/v1"
    description: ProofLink API v1
    variables:
      baseUrl:
        default: "http://localhost:3001"
        description: Base URL of the API server
security:
  - ApiKeyHeader: []
  - BearerAuth: []
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed via X-API-Key header
    BearerAuth:
      type: http
      scheme: bearer
      description: API key or JWT passed via Authorization Bearer header
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              description: Validation error details (array of { path, message })
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    ComplianceParty:
      type: object
      required: [address, chain]
      properties:
        address:
          type: string
          description: Wallet address
        chain:
          type: string
          description: CAIP-2 chain identifier
        agentDID:
          type: string
          description: Optional agent DID
    ComplianceCheckRequest:
      type: object
      required: [sender, receiver, amount, asset]
      properties:
        sender:
          $ref: "#/components/schemas/ComplianceParty"
        receiver:
          $ref: "#/components/schemas/ComplianceParty"
        amount:
          type: string
          description: Transaction amount (decimal string)
        asset:
          type: string
          description: Token symbol or contract address
        protocol:
          type: string
          default: x402
          description: Payment protocol
    CheckPerformed:
      type: object
      properties:
        checkType:
          type: string
          enum:
            - SANCTIONS_SCREENING
            - KYA_VERIFICATION
            - TRAVEL_RULE
            - AML_MONITORING
            - INVOICE_VALIDATION
            - JURISDICTIONAL_RULES
        target:
          type: string
        result:
          type: string
          enum: [PASSED, FAILED, SKIPPED]
        provider:
          type: string
        performedAt:
          type: string
          format: date-time
        durationMs:
          type: integer
    ComplianceCheckResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            status:
              type: string
              enum: [APPROVED, REJECTED, ESCALATED]
            riskScore:
              type: integer
              minimum: 0
              maximum: 100
            receiptId:
              type: string
              format: uuid
            receiptHash:
              type: string
            checks:
              type: array
              items:
                $ref: "#/components/schemas/CheckPerformed"
            travelRuleStatus:
              type: string
              enum: [NOT_REQUIRED, TRANSMITTED, PENDING, FAILED, ACK_RECEIVED]
            totalDurationMs:
              type: integer
            timestamp:
              type: string
              format: date-time
...
    InvoiceLineItem:
      type: object
      required: [description, quantity, unitPrice, total]
      properties:
        description:
          type: string
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          type: string
          default: unit
        unitPrice:
          type: number
          minimum: 0
        total:
          type: number
          minimum: 0
        serviceCategory:
          type: string
          enum:
            - compute
            - data
            - api_call
            - content_generation
            - analysis
            - transaction_fee
            - other
    InvoiceParty:
      type: object
      required: [walletAddress]
      properties:
        agentId:
          type: string
        walletAddress:
          type: string
        legalName:
          type: string
    CreateInvoiceRequest:
      type: object
      required: [seller, buyer, lineItems, currency, totalAmount]
      properties:
        seller:
          $ref: "#/components/schemas/InvoiceParty"
        buyer:
          $ref: "#/components/schemas/InvoiceParty"
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
          minItems: 1
        currency:
          type: string
          enum: [USDC, USDT, USD, EUR, GBP, EURC]
        totalAmount:
          type: number
          minimum: 0
          description: Must match sum of line item totals
        paymentProtocol:
          type: string
          enum: [x402, mpp, ap2, acp, direct]
        dueDate:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
  /compliance/check:
    post:
      tags: [Compliance]
      summary: Run full compliance check
      description: >-
        Runs sanctions screening, KYA verification, AML monitoring, travel rule,
        and jurisdictional checks for a proposed transaction. Returns a
        compliance receipt that can be attached to settlement.
      operationId: complianceCheck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceCheckRequest"
      responses:
        "201":
          description: Compliance check result with receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceCheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /invoices:
    post:
      tags: [Invoices]
      summary: Create a new invoice
      description: >-
        Creates a draft invoice. The totalAmount must match the sum of line item
        totals.
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Invoices]
      summary: List invoices (paginated, filterable)
      operationId: listInvoices
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: state
          in: query
          schema:
            type: string
            enum: [DRAFT, ISSUED, PAID, SETTLED, DISPUTED, CANCELLED]
        - name: currency
          in: query
          schema:
            type: string
            enum: [USDC, USDT, USD, EUR, GBP, EURC]
        - name: seller
          in: query
          schema:
            type: string
          description: Filter by seller wallet address prefix
        - name: buyer
          in: query
          schema:
            type: string
          description: Filter by buyer wallet address prefix
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated invoices
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    const: true
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                      pagination:
                        $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /invoices/{id}:
    get:
      tags: [Invoices]
      summary: Get invoice by ID
      operationId: getInvoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Invoice details
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /analytics/volume:
    get:
      tags: [Analytics]
      summary: Transaction volume over time
      operationId: analyticsVolume
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        "200":
          description: Volume data bucketed by time period
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/compliance:
    get:
      tags: [Analytics]
      summary: Compliance decision breakdown
      operationId: analyticsCompliance
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        "200":
          description: Compliance breakdown with percentages
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/risk:
    get:
      tags: [Analytics]
      summary: Risk score distribution
      operationId: analyticsRisk
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        "200":
          description: Risk distribution data with summary stats
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/agents:
    get:
      tags: [Analytics]
      summary: Top agents by volume
      operationId: analyticsAgents
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Top agents data
        "401":
          $ref: "#/components/responses/Unauthorized"